aes: move legacy RSC loader into its own file
