+++ aes/gemrslib.c
@@ -31,6 +31,7 @@
 #include "geminit.h"
 #include "gemrslib.h"
 #include "gemgsxif.h"
+#include "rsload.h"
 #include "endian.h"
 
 #include "string.h"
@@ -43,30 +44,12 @@
 /* type definitions for use by an application when calling      */
 /*  rsrc_gaddr and rsrc_saddr                                   */
 
-#define R_TREE      0
-#define R_OBJECT    1
-#define R_TEDINFO   2
-#define R_ICONBLK   3
-#define R_BITBLK    4
-#define R_STRING    5               /* gets pointer to free strings */
-#define R_IMAGEDATA 6               /* gets pointer to free images  */
-#define R_OBSPEC    7
-#define R_TEPTEXT   8               /* sub ptrs in TEDINFO  */
-#define R_TEPTMPLT  9
-#define R_TEPVALID  10
-#define R_IBPMASK   11              /* sub ptrs in ICONBLK  */
-#define R_IBPDATA   12
-#define R_IBPTEXT   13
-#define R_BIPDATA   14              /* sub ptrs in BITBLK   */
-#define R_FRSTR     15              /* gets addr of ptr to free strings     */
-#define R_FRIMG     16              /* gets addr of ptr to free images      */
-
 
 
 /*******  LOCALS  **********************/
 
-static RSHDR   *rs_hdr;
-static AESGLOBAL *rs_global;
+RSHDR   *rs_hdr;
+AESGLOBAL *rs_global;
 static char    tmprsfname[128];
 static char    free_str[256];   /* must be long enough for longest freestring in gem.rsc */
 
@@ -499,7 +482,7 @@ void rs_obfix(OBJECT *tree, WORD curob)
 }
 
 
-static void *get_sub(UWORD rsindex, UWORD offset, UWORD rsize)
+void *get_sub(UWORD rsindex, UWORD offset, UWORD rsize)
 {
     /* get base of objects and then index in */
     return (char *)rs_hdr + offset + rsize * rsindex;
@@ -509,7 +492,7 @@ static void *get_sub(UWORD rsindex, UWORD offset, UWORD rsize)
 /*
  *  return address of given type and index, INTERNAL ROUTINE
  */
-static void *get_addr(UWORD rstype, UWORD rsindex)
+void *get_addr(UWORD rstype, UWORD rsindex)
 {
     WORD size;
     UWORD offset;
@@ -577,87 +560,6 @@ static void *get_addr(UWORD rstype, UWORD rsindex)
 
 
 #if CONF_WITH_COLOUR_ICONS
-#if CONF_WITH_LEGACY_RSC_LOAD
-/*
- * fixup all of the CICONs for a CICONBLK
- *
- * returns a pointer to the next CICONBLK
- */
-static CICONBLK *fixup_colour_icons(LONG num_cicons, LONG mono_words, CICON *start)
-{
-    CICON *p = start;
-    WORD *q;
-    LONG i;
-
-    /*
-     * loop through all the CICONs for one CICONBLK
-     *
-     * p points to the current CICON; q tracks sections of CICON data.
-     * at the end, p will point to the start of the next CICONBLK
-     */
-    for (i = 0; i < num_cicons; i++)
-    {
-        q = (WORD *)(p+1);                  /* point to start of data area */
-        p->col_data = q;
-        q += mono_words * p->num_planes;
-        p->col_mask = q;
-        q += mono_words;                    /* mask is one plane */
-        if (p->sel_data)
-        {
-            p->sel_data = q;
-            q += mono_words * p->num_planes;
-            p->sel_mask = q;
-            q += mono_words;                /* mask is one plane */
-        }
-        if (p->next_res == (CICON *)1L)     /* more CICONs follow */
-            p->next_res = (CICON *)q;
-        else
-            p->next_res = NULL;
-        p = (CICON *)q;
-    }
-
-    return (CICONBLK *)p;
-}
-
-/*
- * this fixes up all of the CICONBLK-related pointers:
- *  . the CICONBLK pointer table
- *  . the pointers in the ICONBLK contained in the CICONBLK
- *  . the pointers in the CICON
- */
-static void fixup_all_ciconblks(LONG num_blks, CICONBLK **ciconblkptr, CICON *cicondata)
-{
-    CICONBLK *p = (CICONBLK *)cicondata;
-    WORD *q;
-    LONG i, num_cicons, mono_words;
-
-    for (i = 0; i < num_blks; i++)
-    {
-        ciconblkptr[i] = p;
-        num_cicons = (LONG)(p->mainlist);   /* number of colour icons for this CICONBLK */
-        mono_words = (((LONG)p->monoblk.ib_wicon + 15L) / 16L) * p->monoblk.ib_hicon;
-        q = (WORD *)(p+1);                  /* point to start of data area */
-
-        p->monoblk.ib_pdata = q;            /* fixup mono icon */
-        q += mono_words;
-        p->monoblk.ib_pmask = q;
-        q += mono_words;
-        p->monoblk.ib_ptext = (BYTE *)q;
-        q += 12 / 2;                        /* length of icon text */
-        if (num_cicons)
-        {
-            p->mainlist = (CICON *)q;
-            p = fixup_colour_icons(num_cicons, mono_words, (CICON *)q);
-        }
-        else
-        {
-            p->mainlist = NULL;
-            p = (CICONBLK *)q;
-        }
-    }
-}
-#endif
-
 /*
  * returns pointer to a CICON that best matches the current resolution
  *
@@ -832,7 +734,7 @@ static BOOL pack_cicon(CICON *cicon, WORD w, WORD h)
  * planes that best matches the current resolution.  then expand the icon
  * if necessary, and transform it from standard to device-dependent format
  */
-static void transform_all_cicons(LONG num_cicons, CICONBLK **ciconblkptr)
+void transform_all_cicons(LONG num_cicons, CICONBLK **ciconblkptr)
 {
     CICONBLK *ciconblk;
     CICON *cicon;
@@ -934,7 +836,7 @@ static void transform_all_cicons(LONG num_cicons, CICONBLK **ciconblkptr)
  *
  * returns NULL if none
  */
-static CICONBLK **get_ciconblkptr(RSHDR *hdr)
+CICONBLK **get_ciconblkptr(RSHDR *hdr)
 {
     LONG *extarray;
     LONG cptr_offset;
@@ -999,23 +901,7 @@ static WORD free_cicon_buffers(RSHDR *hdr)
  *      . expanding the icon if necessary
  *      . converting the icon to device-dependent form
  */
-#if CONF_WITH_LEGACY_RSC_LOAD
-static void fix_cicons(void)
-{
-    CICONBLK **ciconblkptr, **p;
-    CICON *cicondata;
-    LONG num_ciconblks;
-
-    ciconblkptr = get_ciconblkptr(rs_hdr);
-    if (!ciconblkptr)
-        return;
-    for (num_ciconblks = 0, p = ciconblkptr; *p != (CICONBLK *)-1L; p++)
-        num_ciconblks++;
-    cicondata = (CICON *)(p+1);
-    fixup_all_ciconblks(num_ciconblks, ciconblkptr, cicondata);
-    transform_all_cicons(num_ciconblks, ciconblkptr);
-}
-#else
+#if !CONF_WITH_LEGACY_RSC_LOAD
 static void transform_cicons(RSHDR *hdr)
 {
     CICONBLK **ciconblkptr, **p;
@@ -1035,70 +921,13 @@ static void transform_cicons(RSHDR *hdr)
 #endif
 
 
-#if CONF_WITH_LEGACY_RSC_LOAD
-static BOOL fix_long(LONG *plong)
-{
-    LONG lngval;
-
-    lngval = *plong;
-    if (lngval != -1L)
-    {
-        *plong = (LONG)rs_hdr + lngval;
-        return TRUE;
-    }
-    return FALSE;
-}
-
-static void fix_trindex(void)
-{
-    WORD ii;
-    LONG *ptreebase;
-
-    ptreebase = (LONG *)get_sub(R_TREE, rs_hdr->rsh_trindex, sizeof(LONG));
-    rs_global->ap_ptree = (OBJECT **)ptreebase;
-    for (ii = 0; ii < rs_hdr->rsh_ntree; ii++)
-        fix_long(ptreebase+ii);
-}
-
-static void fix_nptrs(WORD cnt, WORD type)
-{
-    WORD i;
-
-    for (i = 0; i < cnt; i++)
-        fix_long(get_addr(type, i));
-}
-
-static BOOL fix_ptr(WORD type, WORD index)
-{
-    return fix_long(get_addr(type, index));
-}
-
-static void fix_tedinfo(void)
-{
-    WORD ii;
-    TEDINFO *ted;
-
-    for (ii = 0; ii < rs_hdr->rsh_nted; ii++)
-    {
-        ted = (TEDINFO *)get_addr(R_TEDINFO, ii);
-        if (fix_ptr(R_TEPTEXT, ii))
-            ted->te_txtlen = strlen(ted->te_ptext) + 1;
-        if (fix_ptr(R_TEPTMPLT, ii))
-            ted->te_tmplen = strlen(ted->te_ptmplt) + 1;
-        fix_ptr(R_TEPVALID, ii);
-    }
-}
-#endif
-
-static void fix_objects(void)
+#if !CONF_WITH_LEGACY_RSC_LOAD
+void fix_objects(void)
 {
     WORD ii;
     WORD obtype;
     OBJECT *obj;
 #if CONF_WITH_COLOUR_ICONS
-#if CONF_WITH_LEGACY_RSC_LOAD
-    CICONBLK **ciconblkptr = get_ciconblkptr(rs_hdr);
-#endif
     OBSPEC *spec;
 #endif
 
@@ -1111,18 +940,8 @@ static void fix_objects(void)
         {
         case G_CICON:
 #if CONF_WITH_COLOUR_ICONS
-#if CONF_WITH_LEGACY_RSC_LOAD
-            if (ciconblkptr)
-            {
-                if (obj->ob_flags & INDIRECT)
-                    fix_long(&obj->ob_spec.index);
-                spec = (obj->ob_flags & INDIRECT) ? obj->ob_spec.indirect : &obj->ob_spec;
-                spec->ciconblk = ciconblkptr[spec->index];
-            }
-#else
             spec = (obj->ob_flags & INDIRECT) ? obj->ob_spec.indirect : &obj->ob_spec;
             spec->ciconblk = get_ciconblkptr(rs_hdr)[spec->index];
-#endif
 #endif
             break;
         case G_BOX:
@@ -1130,13 +949,11 @@ static void fix_objects(void)
         case G_BOXCHAR:
             break;
         default:
-#if CONF_WITH_LEGACY_RSC_LOAD
-            fix_long(&obj->ob_spec.index);
-#endif
             break;
         }
     }
 }
+#endif
 
 #if !CONF_WITH_LEGACY_RSC_LOAD
 static void *native_disk_ptr(UBYTE *image, const struct native_rsc_layout *layout,
@@ -1599,57 +1416,8 @@ WORD rs_saddr(AESGLOBAL *pglobal, UWORD rtype, UWORD rindex, void *rsaddr)
  *  case of the GEM resource file this workstation will not have
  *  been loaded into memory yet.
  */
-#if CONF_WITH_LEGACY_RSC_LOAD
-static WORD rs_readit(AESGLOBAL *pglobal, UWORD fd)
-{
-    WORD ibcnt;
-    LONG rslsize;
-    RSHDR hdr_buff;
-
-    if (dos_read(fd, sizeof(hdr_buff), &hdr_buff) != sizeof(hdr_buff))
-        return FALSE;
-    rslsize = hdr_buff.rsh_rssize;
-#if CONF_WITH_COLOUR_ICONS
-    if (hdr_buff.rsh_vrsn & NEW_FORMAT_RSC)
-    {
-        if (dos_lseek(fd, 0, rslsize) < 0L)
-            return FALSE;
-        if (dos_read(fd, sizeof(rslsize), &rslsize) != sizeof(rslsize))
-            return FALSE;
-    }
-#endif
-    rs_hdr = (RSHDR *)dos_alloc_anyram(rslsize);
-    if (!rs_hdr)
-        return FALSE;
-    if (dos_lseek(fd, 0, 0x0L) < 0L)
-        goto fail;
-    if (dos_read(fd, rslsize, rs_hdr) != rslsize)
-        goto fail;
-    rs_global = pglobal;
-    rs_global->ap_rscmem = rs_hdr;
-    rs_global->ap_rsclen = rslsize;
-    fix_trindex();
-#if CONF_WITH_COLOUR_ICONS
-    fix_cicons();
-#endif
-    fix_tedinfo();
-    ibcnt = rs_hdr->rsh_nib;
-    fix_nptrs(ibcnt, R_IBPMASK);
-    fix_nptrs(ibcnt, R_IBPDATA);
-    fix_nptrs(ibcnt, R_IBPTEXT);
-    fix_nptrs(rs_hdr->rsh_nbb, R_BIPDATA);
-    fix_nptrs(rs_hdr->rsh_nstring, R_FRSTR);
-    fix_nptrs(rs_hdr->rsh_nimages, R_FRIMG);
-    return TRUE;
-fail:
-    dos_free(rs_hdr);
-    rs_hdr = NULL;
-    return FALSE;
-}
-#endif
-
 #if !CONF_WITH_LEGACY_RSC_LOAD
-static WORD rs_readit(AESGLOBAL *pglobal,UWORD fd)
+WORD rs_readit(AESGLOBAL *pglobal,UWORD fd)
 {
     UBYTE header[DISK_RSHDR_SIZE];
     struct disk_rsc disk;
